requirements.txt
# Core dependencies
allure-pytest==2.14.0
allure-python-commons==2.14.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Flask==3.0.2
h11==0.14.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==24.2
plotly==5.20.0
pluggy==1.5.0
psutil==7.0.0
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
pyyaml==6.0.2
jsonschema==4.23.0
opencv-python==4.11.0.86
Pillow==11.2.1
scikit-image==0.25.2
matplotlib>=3.7.0
numpy>=1.24.0

# Testing and Coverage
pytest-xdist>=3.3.0
pytest-parallel>=0.1.1
pytest-cov>=3.0.0
pytest-benchmark>=3.4.1
pytest-html>=4.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.7.0

# GraphQL and API
gql>=3.5.0
graphql-core>=3.2.0
aiohttp>=3.9.0

# Monitoring and Visualization
prometheus-client>=0.14.0
grafana-api>=1.0.3
pandas>=2.1.0

# Cloud Integration
boto3>=1.26.0
botocore>=1.29.0
awscli>=1.27.0
google-cloud-compute>=1.10.0
google-cloud-container>=2.30.0
google-cloud-monitoring>=2.15.0
google-cloud-logging>=3.5.0
google-auth>=2.22.0
azure-mgmt-containerservice>=20.0.0
azure-mgmt-resource>=22.0.0
azure-identity>=1.12.0

# Database
psycopg2-binary>=2.9.0
redis>=4.0.0

# Documentation
mkdocs>=1.2.0
mkdocs-material>=8.0.0

# Development Tools
pre-commit>=2.17.0
isort>=5.10.0
mypy>=0.910
flake8>=4.0.0
black>=22.0.0
bandit>=1.7.0

# Notifications and Integration
slack-sdk>=3.26.0
python-telegram-bot>=20.7
pagerduty-api>=2.2.0
pymsteams>=0.2.2

# Performance Testing
locust>=2.8.0

# Blockchain Integration
web3==6.15.1
eth-account==0.11.0
eth-typing==4.0.0
eth-utils==2.4.0
py-solc-x==2.0.0
